# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "examples": [
              "VALIDATION_ERROR"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetailOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "examples": [
              "One or more fields are invalid."
            ],
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "examples": [
              "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
            ],
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetailOut": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetailOut",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PageOut_TaskOut_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "examples": [
              1
            ],
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "examples": [
              20
            ],
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "examples": [
              134
            ],
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "title": "PageOut[TaskOut]",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "StatusChange": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "examples": [
              "in_progress"
            ]
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusChange",
        "type": "object"
      },
      "SummaryOut": {
        "additionalProperties": false,
        "properties": {
          "activeProjects": {
            "minimum": 0.0,
            "title": "Activeprojects",
            "type": "integer"
          },
          "completionRate": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Completionrate",
            "type": "integer"
          },
          "openTasks": {
            "minimum": 0.0,
            "title": "Opentasks",
            "type": "integer"
          },
          "overdueTasks": {
            "minimum": 0.0,
            "title": "Overduetasks",
            "type": "integer"
          },
          "upcomingDeadlines": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "title": "Upcomingdeadlines",
            "type": "array"
          }
        },
        "required": [
          "activeProjects",
          "openTasks",
          "overdueTasks",
          "completionRate",
          "upcomingDeadlines"
        ],
        "title": "SummaryOut",
        "type": "object"
      },
...
      "TaskOut": {
        "additionalProperties": false,
        "properties": {
          "assigneeId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigneeid"
          },
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "description": "UTC, ISO 8601, ends in Z.",
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "dueDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "description": "UTC, ISO 8601, ends in Z.",
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "projectId",
          "title",
          "description",
          "status",
          "priority",
          "dueDate",
          "assigneeId",
          "completedAt"
        ],
        "title": "TaskOut",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "todo",
          "in_progress",
          "in_review",
          "done"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A token from `POST /api/v1/auth/login`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nThe backend for the DevDash dashboard: users, projects, tasks, activity and a dashboard summary.\n\n**Try it in five steps**\n\n1. `POST /api/v1/users` to register (or use a seeded account in development).\n2. `POST /api/v1/auth/login` to get a token.\n3. Click **Authorize** above and paste the `accessToken`.\n4. `POST /api/v1/projects`, then `POST /api/v1/tasks`.\n5. `PATCH /api/v1/tasks/{taskId}/status` to move a task through its workflow.\n\nEvery error has the same shape: `{\"error\": {\"code\", \"message\", \"details\"?, \"requestId\"}}`.\nState is kept in memory and resets on restart (Task 3 adds the database).\n",
    "title": "DevDash API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboard/summary": {
      "get": {
        "description": "Active projects, open tasks, overdue tasks, completion rate, and the open tasks due in the next 7 days, soonest first.",
        "operationId": "summary",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "summary": "UNAUTHENTICATED",
                    "value": {
                      "error": {
                        "code": "UNAUTHENTICATED",
                        "message": "The access token is missing, invalid or expired.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "summary": "INTERNAL_ERROR",
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "Something went wrong on our side.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dashboard summary",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/tasks": {
      "get": {
        "description": "The same filters as `GET /tasks`; the project is fixed by the path.",
        "operationId": "list_project_tasks",
        "parameters": [
          {
            "description": "The project's id.",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "description": "The project's id.",
              "format": "uuid",
              "title": "Projectid",
              "type": "string"
            }
          },
          {
            "description": "Page number, from 1.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number, from 1.",
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page, at most 100.",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page, at most 100.",
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "description": "`field` ascending or `-field` descending. One of: dueDate, priority, title, createdAt.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^-?(dueDate|priority|title|createdAt)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`field` ascending or `-field` descending. One of: dueDate, priority, title, createdAt.",
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "pattern": "^[^\\x00]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/TaskStatus"
              },
              "title": "Status",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/Priority"
              },
              "title": "Priority",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Projectid",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "assigneeId",
            "required": false,
            "schema": {
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Assigneeid",
              "type": "array"
            }
          },
          {
            "description": "BR-04: due before today, not done.",
            "in": "query",
            "name": "overdue",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "BR-04: due before today, not done.",
              "title": "Overdue"
            }
          },
          {
            "in": "query",
            "name": "dueBefore",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duebefore"
            }
          },
          {
            "in": "query",
            "name": "dueAfter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dueafter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut_TaskOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "summary": "UNAUTHENTICATED",
                    "value": {
                      "error": {
                        "code": "UNAUTHENTICATED",
                        "message": "The access token is missing, invalid or expired.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "The resource was not found.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "details": [
                          {
                            "field": "title",
                            "message": "Must be between 1 and 120 characters."
                          }
                        ],
                        "message": "One or more fields are invalid.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "summary": "INTERNAL_ERROR",
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "Something went wrong on our side.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List a project's tasks",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/v1/tasks/{taskId}/status": {
      "patch": {
        "description": "Follows the workflow: todo to in_progress; in_progress to in_review; in_review to in_progress or done; done to in_progress. Anything else is `409 INVALID_STATUS_TRANSITION` and lists the allowed statuses. Requesting the current status succeeds and changes nothing. `completedAt` is set on done and cleared on leaving it.",
        "operationId": "change_status",
        "parameters": [
          {
            "description": "The task's id.",
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "description": "The task's id.",
              "format": "uuid",
              "title": "Taskid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "MALFORMED_REQUEST": {
                    "summary": "MALFORMED_REQUEST",
                    "value": {
                      "error": {
                        "code": "MALFORMED_REQUEST",
                        "message": "The request body is not valid JSON.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed request"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "summary": "UNAUTHENTICATED",
                    "value": {
                      "error": {
                        "code": "UNAUTHENTICATED",
                        "message": "The access token is missing, invalid or expired.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "message": "Only the project owner or a lead may do this.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "The resource was not found.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_STATUS_TRANSITION": {
                    "summary": "INVALID_STATUS_TRANSITION",
                    "value": {
                      "error": {
                        "code": "INVALID_STATUS_TRANSITION",
                        "details": [
                          {
                            "field": "allowedStatuses",
                            "message": "in_progress"
                          }
                        ],
                        "message": "A task cannot move from todo to done.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict with the current state"
          },
          "413": {
            "content": {
              "application/json": {
                "examples": {
                  "PAYLOAD_TOO_LARGE": {
                    "summary": "PAYLOAD_TOO_LARGE",
                    "value": {
                      "error": {
                        "code": "PAYLOAD_TOO_LARGE",
                        "message": "The request body is too large.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "examples": {
                  "UNSUPPORTED_MEDIA_TYPE": {
                    "summary": "UNSUPPORTED_MEDIA_TYPE",
                    "value": {
                      "error": {
                        "code": "UNSUPPORTED_MEDIA_TYPE",
                        "message": "The content type must be application/json.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "details": [
                          {
                            "field": "title",
                            "message": "Must be between 1 and 120 characters."
                          }
                        ],
                        "message": "One or more fields are invalid.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "summary": "INTERNAL_ERROR",
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "Something went wrong on our side.",
                        "requestId": "8f0c2e4a-3b1d-4f6e-9a57-2d7c1e9b5a10"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change a task's status",
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
...
}